+++ src/pages/apps/marketplace/order.vue
@@ -698,7 +698,7 @@ import { bus } from 'src/wallet/event-bus'
 import { backend } from 'src/marketplace/backend'
 import { marketplaceRpc } from 'src/marketplace/rpc'
 import { Delivery, Order, OrderDispute, Payment, Review, Storefront } from 'src/marketplace/objects'
-import { parseCashbackMessage } from 'src/utils/cashback-utils'
+import { parseCashbackMessage } from 'src/utils/engagementhub-utils'
 import { errorParser, formatDateRelative, formatTimestampToText, parsePaymentStatusColor, round } from 'src/marketplace/utils'
 import { debounce, useQuasar } from 'quasar'
 import { useStore } from 'vuex'
